docker: run nanovna-saver from Docker
Create a Dockerfile and a run command to enable running
nanovna-saver from inside a docker image.

This requires a little bit of a dance to allow the GUI access
to the host.
Also, we need to map in the USB device.

Written as my Ubuntu 18.04 system really does not want to run
python3.7 - nanovna-saver will not run on it. What a mess.

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>